third_party/blink/web_tests/virtual/deprecate-unload/external/wpt/permissions-policy/experimental-features/unload-allowed-by-default.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] 'unload' Policy : allowed by default
  assert_equals: sameOriginSubframe: unload in sameOriginSubframe should be allowed expected true but got false
Harness: the test ran to completion.
